% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_badges.R
\name{get_gh_badge}
\alias{get_gh_badge}
\title{Get GitHub badges}
\usage{
get_gh_badge(x)
}
\arguments{
\item{x}{badge}
}
\value{
badges from API
}
\description{
GitHub badges from \href{https://github-profile-summary-cards.vercel.app/demo.html}{vercel}
}
