% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_badges.R
\name{gh_badges}
\alias{gh_badges}
\title{Create GitHub badges}
\usage{
gh_badges(username, badge, theme)
}
\arguments{
\item{username}{GitHub username}

\item{badge}{name of badge ("details", "commit_lang", "ptime", "repo_lang", "stats")}
}
\value{
Markdown links for badges from github-profile-summary-cards.vercel.app API
gh_badges(username = "mjfrigaard", badge = c("details", "stats"), theme = "dark")
}
\description{
GitHub badges from \href{https://github-profile-summary-cards.vercel.app/demo.html}{vercel}
}
