% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghreadme-package.R
\docType{package}
\name{ghreadme-package}
\alias{ghreadme}
\alias{ghreadme-package}
\title{ghreadme: Create a GitHub profile README}
\description{
Quickly create a GitHub profile README from R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mjfrigaard.github.io/ghreadme/}
}

}
\keyword{internal}
