% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so_badges.R
\name{so_rep}
\alias{so_rep}
\title{Stack Overflow Reputation Badge}
\usage{
so_rep(username, user_id)
}
\arguments{
\item{username}{SO username}

\item{user_id}{SO userID}
}
\value{
HTML for badge
}
\description{
Stack Overflow Reputation Badge
}
\examples{
so_rep(username = "martin-frigaard", user_id = "4926446")
}
