% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_am_i.R
\name{who_am_i}
\alias{who_am_i}
\title{Introduce Yourself}
\usage{
who_am_i(name, likes, learn, work, collab, connect = TRUE)
}
\arguments{
\item{name}{Your name}

\item{likes}{Your likes}

\item{learn}{What you're learning}

\item{work}{What you're currently working on}

\item{collab}{What you'd like to collaborate on}

\item{connect}{logical (TRUE/FALSE)}
}
\value{
printed intro
}
\description{
Introduce Yourself
}
\examples{
who_am_i(name = "Martin",
likes = "#rstats and data visualization.",
learn = "shiny app development, JavaScript, and Bayesian statistics.",
work = "R package development tools.",
collab = "#rstats packages for data science.")
}
